Added package support and support for the latest nuget version used with .NET 5
