+++ 50/ConnectingApps.IntegrationFixture50/ConnectingApps.IntegrationFixture50.csproj
@@ -2,13 +2,35 @@
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
+    <IsPackable>true</IsPackable>
+    <Version>5.0.7-dotnet5Support</Version>
+    <OutputType>Dll</OutputType>
+    <Description>An AutoFixture inspired package to create instances for integration testing (not unit testing). Just create an instance of the Fixture class and this library does the work for you to trigger your code in the Startup and Program class. There is a create method to create a test instance and a freeze method to create external dependencies (services) with WireMock.NET .</Description>
+    <PackageLicenseFile>LICENSE</PackageLicenseFile>
+    <RepositoryUrl>https://github.com/ConnectingApps/DncWireMockDemo</RepositoryUrl>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <Copyright>Copyright ©. Connecting Apps. All rights reserved</Copyright>
+    <PackageDescription>An AutoFixture inspired package to create instances for integration testing (not unit testing). Just create an instance of the Fixture class and this library does the work for you to trigger your code in the Startup and Program class. There is a create method to create a test instance and a freeze method to create external dependencies (services) with WireMock.NET </PackageDescription>
+    <PackageProjectUrl>https://github.com/ConnectingApps/DncWireMockDemo#integrationfixture</PackageProjectUrl>
+    <PackageId>ConnectingApps.IntegrationFixture</PackageId>
+    <Authors>Daan Acohen</Authors>
+    <Company>Connecting Apps</Company>
+    <Product>ConnectingApps.IntegrationFixture</Product>
   </PropertyGroup>
 
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.7" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.6.1" />
-    <PackageReference Include="WireMock.Net" Version="1.0.32" />
+    <PackageReference Include="WireMock.Net" Version="1.2.18" />
+  </ItemGroup>
+
+
+  <ItemGroup>
+    <None Include="..\..\LICENSE">
+      <Pack>True</Pack>
+      <PackagePath></PackagePath>
+    </None>
   </ItemGroup>
 
   <Import Project="..\..\ConnectingApps.IntegrationFixture.Shared\ConnectingApps.IntegrationFixture.Shared.projitems" Label="Shared" />
